requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

psycopg2==2.9.9

pillow==10.2.0

redis==5.0.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0
celery==5.3.6
flower==2.0.1

drf-yasg==1.21.7